Split packet.go into multiple files on per-packet-type basis.
