+++ packet-extended-statvfs.go
@@ -1,0 +1,23 @@
+package sftp
+
+type SSHFxpExtendedPacketStatVFS struct {
+	ID              uint32
+	ExtendedRequest string
+	Path            string
+}
+
+func (p SSHFxpExtendedPacketStatVFS) Id() uint32 { return p.ID }
+
+func (p SSHFxpExtendedPacketStatVFS) Readonly() bool { return true }
+
+func (p *SSHFxpExtendedPacketStatVFS) UnmarshalBinary(b []byte) error {
+	var err error
+	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
+		return err
+	} else if p.ExtendedRequest, b, err = unmarshalStringSafe(b); err != nil {
+		return err
+	} else if p.Path, _, err = unmarshalStringSafe(b); err != nil {
+		return err
+	}
+	return nil
+}
