+++ packet.go
@@ -1,9 +1,7 @@
 package sftp
 
 import (
-	"bytes"
 	"encoding"
-	"encoding/binary"
 	"fmt"
 	"io"
 	"os"
@@ -173,76 +171,6 @@ func unmarshalExtensionPair(b []byte) (extensionPair, []byte, error) {
 	return ep, b, err
 }
 
-// Here starts the definition of packets along with their MarshalBinary
-// implementations.
-// Manually writing the marshalling logic wins us a lot of time and
-// allocation.
-
-type SSHFxInitPacket struct {
-	Version    uint32
-	Extensions []extensionPair
-}
-
-// The init packet has no ID, so we just return a zero-value ID
-func (p SSHFxInitPacket) Id() uint32 { return 0 }
-
-func (p SSHFxInitPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 // byte + uint32
-	for _, e := range p.Extensions {
-		l += 4 + len(e.Name) + 4 + len(e.Data)
-	}
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_INIT)
-	b = marshalUint32(b, p.Version)
-	for _, e := range p.Extensions {
-		b = marshalString(b, e.Name)
-		b = marshalString(b, e.Data)
-	}
-	return b, nil
-}
-
-func (p *SSHFxInitPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.Version, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	}
-	for len(b) > 0 {
-		var ep extensionPair
-		ep, b, err = unmarshalExtensionPair(b)
-		if err != nil {
-			return err
-		}
-		p.Extensions = append(p.Extensions, ep)
-	}
-	return nil
-}
-
-type SSHFxVersionPacket struct {
-	Version    uint32
-	Extensions []struct {
-		Name, Data string
-	}
-}
-
-func (p SSHFxVersionPacket) Id() uint32 { return 0 }
-
-func (p SSHFxVersionPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 // byte + uint32
-	for _, e := range p.Extensions {
-		l += 4 + len(e.Name) + 4 + len(e.Data)
-	}
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_VERSION)
-	b = marshalUint32(b, p.Version)
-	for _, e := range p.Extensions {
-		b = marshalString(b, e.Name)
-		b = marshalString(b, e.Data)
-	}
-	return b, nil
-}
-
 func marshalIDString(packetType byte, id uint32, str string) ([]byte, error) {
 	l := 1 + 4 + // type(byte) + uint32
 		4 + len(str)
@@ -264,798 +192,7 @@ func unmarshalIDString(b []byte, id *uint32, str *string) error {
 	return err
 }
 
-type SSHFxpReaddirPacket struct {
-	ID     uint32
-	Handle string
-}
-
-func (p SSHFxpReaddirPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpReaddirPacket) GetHandle() string { return p.Handle }
-
-func (p SSHFxpReaddirPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_READDIR, p.ID, p.Handle)
-}
-
-func (p *SSHFxpReaddirPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Handle)
-}
-
-type SSHFxpOpendirPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpOpendirPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpOpendirPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpOpendirPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_OPENDIR, p.ID, p.Path)
-}
-
-func (p *SSHFxpOpendirPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Path)
-}
-
-type SSHFxpLstatPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpLstatPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpLstatPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpLstatPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_LSTAT, p.ID, p.Path)
-}
-
-func (p *SSHFxpLstatPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Path)
-}
-
-type SSHFxpStatPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpStatPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpStatPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpStatPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_STAT, p.ID, p.Path)
-}
-
-func (p *SSHFxpStatPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Path)
-}
-
-type SSHFxpFstatPacket struct {
-	ID     uint32
-	Handle string
-}
-
-func (p SSHFxpFstatPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpFstatPacket) GetHandle() string { return p.Handle }
-
-func (p SSHFxpFstatPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_FSTAT, p.ID, p.Handle)
-}
-
-func (p *SSHFxpFstatPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Handle)
-}
-
-type SSHFxpClosePacket struct {
-	ID     uint32
-	Handle string
-}
-
-func (p SSHFxpClosePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpClosePacket) GetHandle() string { return p.Handle }
-
-func (p SSHFxpClosePacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_CLOSE, p.ID, p.Handle)
-}
-
-func (p *SSHFxpClosePacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Handle)
-}
-
-type SSHFxpRemovePacket struct {
-	ID       uint32
-	Filename string
-}
-
-func (p SSHFxpRemovePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpRemovePacket) GetPath() string { return p.Filename }
-
-func (p SSHFxpRemovePacket) NotReadOnly() {}
-
-func (p SSHFxpRemovePacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_REMOVE, p.ID, p.Filename)
-}
-
-func (p *SSHFxpRemovePacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Filename)
-}
-
-type SSHFxpRmdirPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpRmdirPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpRmdirPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpRmdirPacket) NotReadOnly() {}
-
-func (p SSHFxpRmdirPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_RMDIR, p.ID, p.Path)
-}
-
-func (p *SSHFxpRmdirPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Path)
-}
-
-type SSHFxpSymlinkPacket struct {
-	ID         uint32
-	Targetpath string
-	Linkpath   string
-}
-
-func (p SSHFxpSymlinkPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpSymlinkPacket) GetPath() string { return p.Targetpath }
-
-func (p SSHFxpSymlinkPacket) NotReadOnly() {}
-
-func (p SSHFxpSymlinkPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Targetpath) +
-		4 + len(p.Linkpath)
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_SYMLINK)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Targetpath)
-	b = marshalString(b, p.Linkpath)
-	return b, nil
-}
-
-func (p *SSHFxpSymlinkPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Targetpath, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Linkpath, _, err = unmarshalStringSafe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-type SSHFxpReadlinkPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpReadlinkPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpReadlinkPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpReadlinkPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_READLINK, p.ID, p.Path)
-}
-
-func (p *SSHFxpReadlinkPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Path)
-}
-
-type SSHFxpRealpathPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpRealpathPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpRealpathPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpRealpathPacket) MarshalBinary() ([]byte, error) {
-	return marshalIDString(ssh_FXP_REALPATH, p.ID, p.Path)
-}
-
-func (p *SSHFxpRealpathPacket) UnmarshalBinary(b []byte) error {
-	return unmarshalIDString(b, &p.ID, &p.Path)
-}
-
-type SSHFxpNameAttr struct {
-	Name     string
-	LongName string
-	Attrs    []interface{}
-}
-
-func (p SSHFxpNameAttr) MarshalBinary() ([]byte, error) {
-	b := []byte{}
-	b = marshalString(b, p.Name)
-	b = marshalString(b, p.LongName)
-	for _, attr := range p.Attrs {
-		b = marshal(b, attr)
-	}
-	return b, nil
-}
-
-type SSHFxpNamePacket struct {
-	ID        uint32
-	NameAttrs []SSHFxpNameAttr
-}
-
-func (p SSHFxpNamePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpNamePacket) MarshalBinary() ([]byte, error) {
-	b := []byte{}
-	b = append(b, ssh_FXP_NAME)
-	b = marshalUint32(b, p.ID)
-	b = marshalUint32(b, uint32(len(p.NameAttrs)))
-	for _, na := range p.NameAttrs {
-		ab, err := na.MarshalBinary()
-		if err != nil {
-			return nil, err
-		}
-
-		b = append(b, ab...)
-	}
-	return b, nil
-}
-
-type SSHFxpOpenPacket struct {
-	ID     uint32
-	Path   string
-	Pflags uint32
-	Flags  uint32 // ignored
-}
-
-func (p SSHFxpOpenPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpOpenPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpOpenPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 +
-		4 + len(p.Path) +
-		4 + 4
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_OPEN)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Path)
-	b = marshalUint32(b, p.Pflags)
-	b = marshalUint32(b, p.Flags)
-	return b, nil
-}
-
-func (p *SSHFxpOpenPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Path, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Pflags, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Flags, _, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (p SSHFxpOpenPacket) Readonly() bool {
-	return !p.HasPflags(ssh_FXF_WRITE)
-}
-
-func (p SSHFxpOpenPacket) HasPflags(flags ...uint32) bool {
-	for _, f := range flags {
-		if p.Pflags&f == 0 {
-			return false
-		}
-	}
-	return true
-}
-
-type SSHFxpReadPacket struct {
-	ID     uint32
-	Handle string
-	Offset uint64
-	Len    uint32
-}
-
-func (p SSHFxpReadPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpReadPacket) GetHandle() string { return p.Handle }
-
-func (p SSHFxpReadPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Handle) +
-		8 + 4 // uint64 + uint32
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_READ)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Handle)
-	b = marshalUint64(b, p.Offset)
-	b = marshalUint32(b, p.Len)
-	return b, nil
-}
-
-func (p *SSHFxpReadPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Handle, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Offset, b, err = unmarshalUint64Safe(b); err != nil {
-		return err
-	} else if p.Len, _, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-type SSHFxpRenamePacket struct {
-	ID      uint32
-	Oldpath string
-	Newpath string
-}
-
-func (p SSHFxpRenamePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpRenamePacket) GetPath() string { return p.Oldpath }
-
-func (p SSHFxpRenamePacket) NotReadOnly() {}
-
-func (p SSHFxpRenamePacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Oldpath) +
-		4 + len(p.Newpath)
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_RENAME)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Oldpath)
-	b = marshalString(b, p.Newpath)
-	return b, nil
-}
-
-func (p *SSHFxpRenamePacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Oldpath, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Newpath, _, err = unmarshalStringSafe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-type SSHFxpPosixRenamePacket struct {
-	ID      uint32
-	Oldpath string
-	Newpath string
-}
-
-func (p SSHFxpPosixRenamePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpPosixRenamePacket) GetPath() string { return p.Oldpath }
-
-func (p SSHFxpPosixRenamePacket) Readonly() bool { return false }
-
-func (p SSHFxpPosixRenamePacket) NotReadOnly() {}
-
-func (p SSHFxpPosixRenamePacket) MarshalBinary() ([]byte, error) {
-	const ext = "posix-rename@openssh.com"
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(ext) +
-		4 + len(p.Oldpath) +
-		4 + len(p.Newpath)
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_EXTENDED)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, ext)
-	b = marshalString(b, p.Oldpath)
-	b = marshalString(b, p.Newpath)
-	return b, nil
-}
-
-type SSHFxpWritePacket struct {
-	ID     uint32
-	Handle string
-	Offset uint64
-	Length uint32
-	Data   []byte
-}
-
-func (p SSHFxpWritePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpWritePacket) GetHandle() string { return p.Handle }
-
-func (p SSHFxpWritePacket) NotReadOnly() {}
-
-func (p SSHFxpWritePacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Handle) +
-		8 + 4 + // uint64 + uint32
-		len(p.Data)
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_WRITE)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Handle)
-	b = marshalUint64(b, p.Offset)
-	b = marshalUint32(b, p.Length)
-	b = append(b, p.Data...)
-	return b, nil
-}
-
-func (p *SSHFxpWritePacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Handle, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Offset, b, err = unmarshalUint64Safe(b); err != nil {
-		return err
-	} else if p.Length, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if uint32(len(b)) < p.Length {
-		return errShortPacket
-	}
-
-	p.Data = append([]byte{}, b[:p.Length]...)
-	return nil
-}
-
-type SSHFxpMkdirPacket struct {
-	ID    uint32
-	Path  string
-	Flags uint32 // ignored
-}
-
-func (p SSHFxpMkdirPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpMkdirPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpMkdirPacket) NotReadOnly() {}
-
-func (p SSHFxpMkdirPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Path) +
-		4 // uint32
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_MKDIR)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Path)
-	b = marshalUint32(b, p.Flags)
-	return b, nil
-}
-
-func (p *SSHFxpMkdirPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Path, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Flags, _, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-type SSHFxpSetstatPacket struct {
-	ID    uint32
-	Path  string
-	Flags uint32
-	Attrs interface{}
-}
-
-func (p SSHFxpSetstatPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpSetstatPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpSetstatPacket) NotReadOnly() {}
-
-func (p SSHFxpSetstatPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Path) +
-		4 // uint32 + uint64
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_SETSTAT)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Path)
-	b = marshalUint32(b, p.Flags)
-	b = marshal(b, p.Attrs)
-	return b, nil
-}
-
-func (p *SSHFxpSetstatPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Path, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Flags, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	}
-	p.Attrs = b
-	return nil
-}
-
-type SSHFxpFsetstatPacket struct {
-	ID     uint32
-	Handle string
-	Flags  uint32
-	Attrs  interface{}
-}
-
-func (p SSHFxpFsetstatPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpFsetstatPacket) GetHandle() string { return p.Handle }
-
-func (p SSHFxpFsetstatPacket) NotReadOnly() {}
-
-func (p SSHFxpFsetstatPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		4 + len(p.Handle) +
-		4 // uint32 + uint64
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_FSETSTAT)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Handle)
-	b = marshalUint32(b, p.Flags)
-	b = marshal(b, p.Attrs)
-	return b, nil
-}
-
-func (p *SSHFxpFsetstatPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Handle, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Flags, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	}
-	p.Attrs = b
-	return nil
-}
-
-type SSHFxpHandlePacket struct {
-	ID     uint32
-	Handle string
-}
-
-func (p SSHFxpHandlePacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpHandlePacket) MarshalBinary() ([]byte, error) {
-	b := []byte{ssh_FXP_HANDLE}
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, p.Handle)
-	return b, nil
-}
-
-type SSHFxpStatusPacket struct {
-	ID uint32
-	StatusError
-}
-
-func (p SSHFxpStatusPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpStatusPacket) MarshalBinary() ([]byte, error) {
-	b := []byte{ssh_FXP_STATUS}
-	b = marshalUint32(b, p.ID)
-	b = marshalStatus(b, p.StatusError)
-	return b, nil
-}
-
-type SSHFxpDataPacket struct {
-	ID     uint32
-	Length uint32
-	Data   []byte
-}
-
-func (p SSHFxpDataPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpDataPacket) MarshalBinary() ([]byte, error) {
-	b := []byte{ssh_FXP_DATA}
-	b = marshalUint32(b, p.ID)
-	b = marshalUint32(b, p.Length)
-	b = append(b, p.Data[:p.Length]...)
-	return b, nil
-}
-
-func (p *SSHFxpDataPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.Length, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if uint32(len(b)) < p.Length {
-		return errors.New("truncated packet")
-	}
-
-	p.Data = make([]byte, p.Length)
-	copy(p.Data, b)
-	return nil
-}
-
-type SSHFxpStatvfsPacket struct {
-	ID   uint32
-	Path string
-}
-
-func (p SSHFxpStatvfsPacket) Id() uint32 { return p.ID }
-
-func (p SSHFxpStatvfsPacket) GetPath() string { return p.Path }
-
-func (p SSHFxpStatvfsPacket) MarshalBinary() ([]byte, error) {
-	l := 1 + 4 + // type(byte) + uint32
-		len(p.Path) +
-		len("statvfs@openssh.com")
-
-	b := make([]byte, 0, l)
-	b = append(b, ssh_FXP_EXTENDED)
-	b = marshalUint32(b, p.ID)
-	b = marshalString(b, "statvfs@openssh.com")
-	b = marshalString(b, p.Path)
-	return b, nil
-}
-
-// A StatVFS contains statistics about a filesystem.
-type StatVFS struct {
-	ID      uint32
-	Bsize   uint64 /* file system block size */
-	Frsize  uint64 /* fundamental fs block size */
-	Blocks  uint64 /* number of blocks (unit f_frsize) */
-	Bfree   uint64 /* free blocks in file system */
-	Bavail  uint64 /* free blocks for non-root */
-	Files   uint64 /* total file inodes */
-	Ffree   uint64 /* free file inodes */
-	Favail  uint64 /* free file inodes for to non-root */
-	Fsid    uint64 /* file system id */
-	Flag    uint64 /* bit mask of f_flag values */
-	Namemax uint64 /* maximum filename length */
-}
-
-func (p *StatVFS) Id() uint32 {
-	return p.ID
-}
-
-// TotalSpace calculates the amount of total space in a filesystem.
-func (p *StatVFS) TotalSpace() uint64 {
-	return p.Frsize * p.Blocks
-}
-
-// FreeSpace calculates the amount of free space in a filesystem.
-func (p *StatVFS) FreeSpace() uint64 {
-	return p.Frsize * p.Bfree
-}
-
-// Convert to ssh_FXP_EXTENDED_REPLY packet binary format
-func (p *StatVFS) MarshalBinary() ([]byte, error) {
-	var buf bytes.Buffer
-	buf.Write([]byte{ssh_FXP_EXTENDED_REPLY})
-	err := binary.Write(&buf, binary.BigEndian, p)
-	return buf.Bytes(), err
-}
-
-type SSHFxpExtendedPacket struct {
-	ID              uint32
-	ExtendedRequest string
-	SpecificPacket  interface {
-		RequestPacket
-		Readonly() bool
-	}
-}
-
-func (p SSHFxpExtendedPacket) Id() uint32 { return p.ID }
-func (p SSHFxpExtendedPacket) Readonly() bool {
-	if p.SpecificPacket == nil {
-		return true
-	}
-	return p.SpecificPacket.Readonly()
-}
-
-func (p *SSHFxpExtendedPacket) UnmarshalBinary(b []byte) error {
-	var err error
-	bOrig := b
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.ExtendedRequest, _, err = unmarshalStringSafe(b); err != nil {
-		return err
-	}
-
-	// specific unmarshalling
-	switch p.ExtendedRequest {
-	case "statvfs@openssh.com":
-		p.SpecificPacket = &SSHFxpExtendedPacketStatVFS{}
-	case "posix-rename@openssh.com":
-		p.SpecificPacket = &SSHFxpExtendedPacketPosixRename{}
-	default:
-		return errors.Wrapf(errUnknownExtendedPacket, "packet type %v", p.SpecificPacket)
-	}
-
-	return p.SpecificPacket.UnmarshalBinary(bOrig)
-}
-
-type SSHFxpExtendedPacketStatVFS struct {
-	ID              uint32
-	ExtendedRequest string
-	Path            string
-}
-
-func (p SSHFxpExtendedPacketStatVFS) Id() uint32 { return p.ID }
-
-func (p SSHFxpExtendedPacketStatVFS) Readonly() bool { return true }
-
-func (p *SSHFxpExtendedPacketStatVFS) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.ExtendedRequest, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Path, _, err = unmarshalStringSafe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-type SSHFxpExtendedPacketPosixRename struct {
-	ID              uint32
-	ExtendedRequest string
-	Oldpath         string
-	Newpath         string
-}
-
-func (p SSHFxpExtendedPacketPosixRename) Id() uint32 { return p.ID }
-
-func (p SSHFxpExtendedPacketPosixRename) GetPath() string { return p.Oldpath }
-
-func (p SSHFxpExtendedPacketPosixRename) Readonly() bool { return false }
-
-func (p SSHFxpExtendedPacketPosixRename) NotReadOnly() {}
-
-func (p *SSHFxpExtendedPacketPosixRename) UnmarshalBinary(b []byte) error {
-	var err error
-	if p.ID, b, err = unmarshalUint32Safe(b); err != nil {
-		return err
-	} else if p.ExtendedRequest, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Oldpath, b, err = unmarshalStringSafe(b); err != nil {
-		return err
-	} else if p.Newpath, _, err = unmarshalStringSafe(b); err != nil {
-		return err
-	}
-	return nil
-}
-
-type SSHFxpStatResponse struct {
-	ID   uint32
-	Info os.FileInfo
-}
-
-func (p SSHFxpStatResponse) Id() uint32 { return p.ID }
-
-func (p SSHFxpStatResponse) MarshalBinary() ([]byte, error) {
-	b := []byte{ssh_FXP_ATTRS}
-	b = marshalUint32(b, p.ID)
-	b = marshalFileInfo(b, p.Info)
-	return b, nil
-}
+// Here starts the definition of packets along with their MarshalBinary
+// implementations.
+// Manually writing the marshalling logic wins us a lot of time and
+// allocation.
